refactor: extract shared handlebars Vite plugin
vite.config.ts and vitest.config.ts each had their own inline copy
of a handlebars .hbs loader. The two had already drifted — the Vite
copy escaped <, >, / and the line/paragraph separators; the Vitest
copy only escaped the separators. Extracting both to
build-utils/handlebars-loader.ts means they stay in sync and the
safer escape set is used everywhere.
