+++ vitest.config.ts
@@ -1,27 +1,9 @@
 import { defineConfig } from 'vitest/config';
 import { resolve } from 'path';
-import { readFileSync } from 'fs';
-
-function escapeUnsafeChars(str: string): string {
-  return str.replace(/\u2028/g, '\\u2028').replace(/\u2029/g, '\\u2029');
-}
+import { handlebarsLoader } from './build-utils/handlebars-loader';
 
 export default defineConfig({
-  plugins: [
-    {
-      name: 'handlebars-loader',
-      transform(_code: string, id: string) {
-        if (id.endsWith('.hbs')) {
-          const template = readFileSync(id, 'utf-8');
-          return {
-            code: `export default function() { return ${escapeUnsafeChars(JSON.stringify(template))}; }`,
-            map: null,
-          };
-        }
-        return null;
-      },
-    },
-  ],
+  plugins: [handlebarsLoader()],
   test: {
     environment: 'jsdom',
     globals: true,
